+++ virtual_tcu/input/__init__.py
@@ -1,6 +1,5 @@
-from virtual_tcu.input.gamepad_output import GamepadOutput
 from virtual_tcu.input.interface import OutputInterface
 from virtual_tcu.input.keyboard_output import KeyboardOutput
 from virtual_tcu.input.vjoy_output import VJoyOutput
 
-__all__ = ["GamepadOutput", "KeyboardOutput", "VJoyOutput", "OutputInterface"]
+__all__ = ["KeyboardOutput", "VJoyOutput", "OutputInterface"]